+++ src/Ciromattia/Teamwork/Contracts/RequestableInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace Ciromattia\Teamwork\Contracts;
+
+interface RequestableInterface
+{
+
+}