+++ src/Ciromattia/Teamwork/Exceptions/ClassNotCreatedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Ciromattia\Teamwork\Exceptions;
+
+class ClassNotCreatedException extends \Exception
+{
+}