+++ src/Ciromattia/Teamwork/Tasklist.php
@@ -1,11 +1,11 @@
-<?php  namespace Rossedman\Teamwork;
+<?php
 
-use Rossedman\Teamwork\Traits\RestfulTrait;
+namespace Ciromattia\Teamwork;
 
-class Tasklist extends AbstractObject {
+class Tasklist extends AbstractObject
+{
 
-    protected $wrapper  = 'todo-list';
-
+    protected $wrapper = 'todo-list';
     protected $endpoint = 'tasklists';
 
     /**
@@ -66,7 +66,7 @@ class Tasklist extends AbstractObject {
     public function tasks($args = null)
     {
         $this->areArgumentsValid($args, ['filter', 'page', 'pageSize', 'startdate', 'enddate', 'updatedAfterDate', 'completedAfterDate', 'completedBeforeDate', 'showDeleted', 'includeCompletedTasks', 'includeCompletedSubtasks', 'creator-ids', 'include', 'responsible-party-ids', 'sort', 'getSubTasks', 'nestSubTasks', 'getFiles', 'dataSet', 'includeToday', 'ignore-start-date']);
-    
+
         return $this->client->get("$this->endpoint/$this->id/tasks", $args)->response();
     }