+++ changes.md
@@ -1,6 +1,6 @@
 ## v1.0.1
 * Update docs
-* Create documtentation website
+* Create documentation website
 * Add createTasklist and createMilestone methods
 
 ## v1.0.0